some fixes and optimisations
