+++ src/components/MemberSocials/MemberSocials.module.css
@@ -1,0 +1,12 @@
+.contact-btn {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	background-color: #343434;
+	border-radius: 5px;
+	height: 45px;
+}
+
+.contact-btn svg {
+	fill: rgba(255, 255, 255, 0.6)
+}
